requirements.txt
fastapi
uvicorn
streamlit
ffmpeg-python
openai-whisper
requests

scenedetect
opencv-python-headless
